canvas2d/fill-and-stroke-styles/2d.pattern.basic.zerocanvas.html

<!DOCTYPE html>
<title>Canvas test: 2d.pattern.basic.zerocanvas</title>
<meta name="author" content="Philip Taylor">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../common/canvas-tests.js"></script>
<link rel="stylesheet" href="../../../common/canvas-tests.css">
<body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.pattern.html">pattern</a>.<a href="index.2d.pattern.basic.html">basic</a>.zerocanvas</h1>
<p class="desc"></p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.pattern.zerocanvas">2d.pattern.zerocanvas</a>
</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

canvas.width = 0;
canvas.height = 10;
_assertSame(canvas.width, 0, "canvas.width", "0");
_assertSame(canvas.height, 10, "canvas.height", "10");
assert_throws("INVALID_STATE_ERR", function() { ctx.createPattern(canvas, 'repeat'); });

canvas.width = 10;
canvas.height = 0;
_assertSame(canvas.width, 10, "canvas.width", "10");
_assertSame(canvas.height, 0, "canvas.height", "0");
assert_throws("INVALID_STATE_ERR", function() { ctx.createPattern(canvas, 'repeat'); });

canvas.width = 0;
canvas.height = 0;
_assertSame(canvas.width, 0, "canvas.width", "0");
_assertSame(canvas.height, 0, "canvas.height", "0");
assert_throws("INVALID_STATE_ERR", function() { ctx.createPattern(canvas, 'repeat'); });


});
</script>
